Update D&D 5E Macro Mule
Update 9/19/22:
I've adjusted the Collection Macros 'Visible To' and 'Is Token Action' settings.

I've also added a new feature along with the updates for PC/NPC whisper settings.  As long as you have a Roll20 Pro subscription, you can use a 'whisper to myself' command that I've added to this script, which is simply:

!self <message to send to yourself>
This command works for both players and the GM, and it's the command that is created for the 'NPCWhisper' and 'PCWhisper' attributes.  If you ever drop your Roll20 Pro subscription, you can change those two attributes to the manual approach that I've listed for non-Pro subscribers.